fix(template): implemented partial abstraction that were moved to test/index.ts
